requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
fastapi-admin
python-dotenv
pandas
python-jose
passlib[bcrypt]
